did some refactoring, MVP is done for homepage, or at least 95% done.
