+++ app/src/main/java/com/example/cloudcards/Homepage.java
@@ -2,64 +2,48 @@ package com.example.cloudcards;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 import android.Manifest;
 import android.content.ContentValues;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.util.Log;
-import android.util.SparseArray;
 import android.view.ContextThemeWrapper;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.PopupMenu;
 import android.widget.Toast;
-import com.example.cloudcards.database.DBHelper;
-import com.google.android.gms.vision.Frame;
-import com.google.android.gms.vision.text.TextBlock;
-import com.google.android.gms.vision.text.TextRecognizer;
+import com.example.cloudcards.Presenter.HomepageActivityPresenter;
 import com.theartofdev.edmodo.cropper.CropImage;
 import com.theartofdev.edmodo.cropper.CropImageView;
-import org.json.JSONObject;
-import java.io.IOException;
-import java.util.ArrayList;
 
-
-public class Homepage extends AppCompatActivity  implements MenuItem.OnMenuItemClickListener {
+public class Homepage extends AppCompatActivity implements MenuItem.OnMenuItemClickListener {
 
     Button showMenu;
     private static final int REQUEST_IMAGE_CAPTURE = 1;
-    private static final int SEARCH_ACTIVITY_REQUEST_CODE = 2;
     Uri image_uri;
     String cameraPermission[];
     String storagePermission[];
-    String name;
     int userID;
-    private APIHelper API;
-    private DBHelper DB;
+    private HomepageActivityPresenter presenter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.homepage);
-        // Setup card collection
+        // user ID
         userID = getIntent().getIntExtra("userID", 0);
-        API = new APIHelper(getApplicationContext());
-        DB = new DBHelper(getApplicationContext());
         // permissions
         cameraPermission = new String[]{Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE};
         storagePermission = new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE};
-        showMenu = (Button) findViewById(R.id.backButton);
+        presenter = new HomepageActivityPresenter(getApplicationContext(), this, userID);
+        showMenu = (Button) findViewById(R.id.menu);
         showMenu.setText("Menu");
         showMenu.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -71,7 +55,7 @@ public class Homepage extends AppCompatActivity  implements MenuItem.OnMenuItemC
                     @Override
                     public boolean onMenuItemClick(MenuItem menuItem) {
                         switch (menuItem.getTitle().toString()) {
-                            case "Add Card": takePhoto();
+                             case "Add Card": takePhoto();
                             break;
                             case "View Collection":
                                 Intent intent = new Intent(getApplicationContext(), Collection.class);
@@ -113,41 +97,14 @@ public class Homepage extends AppCompatActivity  implements MenuItem.OnMenuItemC
 
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == SEARCH_ACTIVITY_REQUEST_CODE) {
-        }
         if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
             CropImage.activity(image_uri).setGuidelines(CropImageView.Guidelines.ON).start(this);
         }
         if (requestCode == CropImage.CROP_IMAGE_ACTIVITY_REQUEST_CODE){
             CropImage.ActivityResult result = CropImage.getActivityResult(data);
             if (resultCode == RESULT_OK) {
-                Uri resultUri = result.getUri();
-                // get drawable bitmap for text recognition
-                Bitmap uriBit =  null;
-                try {
-                    uriBit = MediaStore.Images.Media.getBitmap(this.getContentResolver(), resultUri);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                TextRecognizer recognizer = new TextRecognizer.Builder(getApplicationContext()).build();
-                if (!recognizer.isOperational()) {
-                    Toast.makeText(this, "Error", Toast.LENGTH_SHORT).show();
-                } else {
-                    Frame frame = new Frame.Builder().setBitmap(uriBit).build();
-                    SparseArray<TextBlock> items = recognizer.detect(frame);
-                    name = items.valueAt(0).getValue();
-                    API.getCardByName(name, new VolleyCallback() {
-                        @Override
-                        public void onSuccess(JSONObject result) {
-                            if (result != null) {
-                                Card card = new Card(result);
-                                dialogueAddCard(card);
-                            } else {
-                                Toast.makeText(getApplicationContext(), "Unable to Identify Card, please try again.", Toast.LENGTH_LONG).show();
-                            }
-                        }
-                    });
-                }
+                // Presenter does card recognition, API call, and dialogue.
+                presenter.detectTextOnCropResult(result);
             } else if (resultCode == CropImage.CROP_IMAGE_ACTIVITY_RESULT_ERROR_CODE){
                 Exception error = result.getError();
                 Toast.makeText(this, ""+error, Toast.LENGTH_SHORT).show();
@@ -186,39 +143,5 @@ public class Homepage extends AppCompatActivity  implements MenuItem.OnMenuItemC
                 }
         }
     }
-
-
-    private void dialogueAddCard(Card card){
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setMessage("Do you want to add this card to your collection?"+"\n"+card.getCard_name());
-        builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-                        boolean result = dbAddCard(card);
-                        if (result == true) {
-                            Toast.makeText(getApplicationContext(), "Card added to collection.", Toast.LENGTH_LONG).show();
-                        } else {
-                            Toast.makeText(getApplicationContext(), "Add Card Failed.", Toast.LENGTH_LONG).show();
-                        }
-                    }
-                });
-        builder.setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-                        dialogInterface.cancel();
-                        Toast.makeText(getApplicationContext(), "Card not added to Collection", Toast.LENGTH_LONG).show();
-                    }
-                });
-        builder.setTitle("Confirm Card");
-        builder.show();
-    }
-
-    private boolean dbAddCard(Card card){
-        boolean result = DB.insertCard(userID, card);
-        ArrayList<Card> cards = DB.getCardsByUserID(userID);
-        Log.i("cards", cards.toString());
-        return result;
-    }
-
 
 }
