+++ app/src/main/res/layout/homepage.xml
@@ -8,7 +8,7 @@
     tools:context=".Homepage">
 
     <Button
-        android:id="@+id/backButton"
+        android:id="@+id/menu"
         android:layout_width="88dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="12dp"
